Fix issue with implicit heat filter - wrong usage of time steps number

+++ ImageInLib/ImageInLib/src/filter_params.h
@@ -6,7 +6,7 @@ extern "C" {
 #pragma once
 	typedef struct
 	{
-		size_t timeStepSize;
+		double timeStepSize;
 		double h;
 		double sigma;
 		double edge_detector_coefficient;
